renamed "core.levels" into "core.level" renamed "core.api" into "core.xarray"
