+++ test/core/test_levels.py
@@ -5,7 +5,7 @@ import unittest
 import numpy as np
 import xarray as xr
 
-from xcube.core.levels import compute_levels, write_levels, read_levels
+from xcube.core.level import compute_levels, write_levels, read_levels
 from xcube.util.dsio import rimraf
